requirements.txt
# ==============================================
# Pillow 智能枕头项目 - Python 依赖
# ==============================================

# ----------------------------------------------
# 核心依赖 (所有环境必需)
# ----------------------------------------------
numpy>=1.21.0
librosa>=0.9.0
tqdm>=4.60.0

# ----------------------------------------------
# 深度学习 (训练 & PyTorch推理)
# ----------------------------------------------
torch>=1.10.0
# torchvision  # 如需图像处理可取消注释

# ----------------------------------------------
# 实验追踪 (训练时需要)
# ----------------------------------------------
wandb>=0.12.0

# ----------------------------------------------
# 音频采集 (实时推理需要)
# ----------------------------------------------
sounddevice>=0.4.0

# ----------------------------------------------
# 硬件控制 (气囊控制器)
# ----------------------------------------------
pyserial>=3.5
pynput>=1.7.0

# ==============================================
# 可选依赖 (按需安装)
# ==============================================

# RKNN NPU 推理 (仅 RK3588/RK3566 等 Rockchip 开发板)
# 需要从官方获取对应版本的 wheel 包:
# https://github.com/airockchip/rknn-toolkit2
# pip install rknn-toolkit-lite2-x.x.x-cpxx-cpxx-linux_aarch64.whl
# rknn-toolkit-lite2

# ONNX 导出 (模型转换时需要)
# onnx>=1.10.0
# onnxruntime>=1.10.0
